Refactor ParserPath into its own file
