+++ crates/nu-parser/src/lib.rs
@@ -7,6 +7,7 @@ mod lite_parser;
 mod parse_keywords;
 mod parse_patterns;
 mod parser;
+mod parser_path;
 mod type_check;
 
 pub use deparse::{escape_for_script_arg, escape_quote_string};
@@ -17,6 +18,7 @@ pub use known_external::KnownExternal;
 pub use lex::{lex, lex_signature, Token, TokenContents};
 pub use lite_parser::{lite_parse, LiteBlock, LiteElement};
 pub use parse_keywords::*;
+pub use parser_path::*;
 
 pub use parser::{
     is_math_expression_like, parse, parse_block, parse_expression, parse_external_call,
