+++ crates/nu-parser/src/parse_keywords.rs
@@ -1,3 +1,4 @@
+use crate::parser_path::ParserPath;
 use itertools::Itertools;
 use log::trace;
 use nu_path::canonicalize_with;
@@ -6,7 +7,7 @@ use nu_protocol::{
         Argument, Block, Call, Expr, Expression, ImportPattern, ImportPatternHead,
         ImportPatternMember, Pipeline, PipelineElement,
     },
-    engine::{StateWorkingSet, VirtualPath, DEFAULT_OVERLAY_NAME},
+    engine::{StateWorkingSet, DEFAULT_OVERLAY_NAME},
     span, Alias, BlockId, Exportable, Module, ModuleId, ParseError, PositionalArg,
     ResolvedImportPattern, Span, Spanned, SyntaxShape, Type, VarId,
 };
@@ -3351,132 +3352,6 @@ pub fn parse_register(working_set: &mut StateWorkingSet, spans: &[Span]) -> Pipe
         ty: Type::Nothing,
         custom_completion: None,
     }])
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub enum ParserPath {
-    RealPath(PathBuf),
-    VirtualFile(PathBuf, usize),
-    VirtualDir(PathBuf, Vec<ParserPath>),
-}
-
-impl ParserPath {
-    pub fn is_dir(&self) -> bool {
-        match self {
-            ParserPath::RealPath(p) => p.is_dir(),
-            ParserPath::VirtualFile(..) => false,
-            ParserPath::VirtualDir(..) => true,
-        }
-    }
-
-    pub fn is_file(&self) -> bool {
-        match self {
-            ParserPath::RealPath(p) => p.is_file(),
-            ParserPath::VirtualFile(..) => true,
-            ParserPath::VirtualDir(..) => false,
-        }
-    }
-
-    pub fn exists(&self) -> bool {
-        match self {
-            ParserPath::RealPath(p) => p.exists(),
-            ParserPath::VirtualFile(..) => true,
-            ParserPath::VirtualDir(..) => true,
-        }
-    }
-
-    pub fn path(&self) -> &Path {
-        match self {
-            ParserPath::RealPath(p) => p,
-            ParserPath::VirtualFile(p, _) => p,
-            ParserPath::VirtualDir(p, _) => p,
-        }
-    }
-
-    pub fn path_buf(self) -> PathBuf {
-        match self {
-            ParserPath::RealPath(p) => p,
-            ParserPath::VirtualFile(p, _) => p,
-            ParserPath::VirtualDir(p, _) => p,
-        }
-    }
-
-    pub fn parent(&self) -> Option<&Path> {
-        match self {
-            ParserPath::RealPath(p) => p.parent(),
-            ParserPath::VirtualFile(p, _) => p.parent(),
-            ParserPath::VirtualDir(p, _) => p.parent(),
-        }
-    }
-
-    pub fn read_dir(&self) -> Option<Vec<ParserPath>> {
-        match self {
-            ParserPath::RealPath(p) => p.read_dir().ok().map(|read_dir| {
-                read_dir
-                    .flatten()
-                    .map(|dir_entry| ParserPath::RealPath(dir_entry.path()))
-                    .collect()
-            }),
-            ParserPath::VirtualFile(..) => None,
-            ParserPath::VirtualDir(_, files) => Some(files.clone()),
-        }
-    }
-
-    pub fn file_stem(&self) -> Option<&OsStr> {
-        self.path().file_stem()
-    }
-
-    pub fn extension(&self) -> Option<&OsStr> {
-        self.path().extension()
-    }
-
-    pub fn join(self, path: impl AsRef<Path>) -> ParserPath {
-        match self {
-            ParserPath::RealPath(p) => ParserPath::RealPath(p.join(path)),
-            ParserPath::VirtualFile(p, file_id) => ParserPath::VirtualFile(p.join(path), file_id),
-            ParserPath::VirtualDir(p, entries) => {
-                let new_p = p.join(path);
-                let mut pp = ParserPath::RealPath(new_p.clone());
-                for entry in entries {
-                    if new_p == entry.path() {
-                        pp = entry.clone();
-                    }
-                }
-                pp
-            }
-        }
-    }
-
-    pub fn read<'a>(&'a self, working_set: &'a StateWorkingSet) -> Option<Vec<u8>> {
-        match self {
-            ParserPath::RealPath(p) => std::fs::read(p).ok(),
-            ParserPath::VirtualFile(_, file_id) => working_set
-                .get_contents_of_file(*file_id)
-                .map(|bytes| bytes.to_vec()),
-
-            ParserPath::VirtualDir(..) => None,
-        }
-    }
-
-    pub fn from_virtual_path(
-        working_set: &StateWorkingSet,
-        name: &str,
-        virtual_path: &VirtualPath,
-    ) -> Self {
-        match virtual_path {
-            VirtualPath::File(file_id) => ParserPath::VirtualFile(PathBuf::from(name), *file_id),
-            VirtualPath::Dir(entries) => ParserPath::VirtualDir(
-                PathBuf::from(name),
-                entries
-                    .iter()
-                    .map(|virtual_path_id| {
-                        let (virt_name, virt_path) = working_set.get_virtual_path(*virtual_path_id);
-                        ParserPath::from_virtual_path(working_set, virt_name, virt_path)
-                    })
-                    .collect(),
-            ),
-        }
-    }
 }
 
 pub fn find_dirs_var(working_set: &StateWorkingSet, var_name: &str) -> Option<VarId> {
